Cleanup proxy setup
