+++ api/python/cellxgene_census/src/cellxgene_census/_testing/logger_proxy.py
@@ -1,0 +1,39 @@
+"""This module defines a plugin class that logs each request to the logfile."""
+
+import json
+import traceback
+from pathlib import Path
+
+import proxy
+from proxy.common.flag import flags
+
+flags.add_argument(
+    "--request-logfile",
+    type=str,
+    default="",
+    help="Where to log the requests to.",
+)
+
+
+class RequestLoggerPlugin(proxy.http.proxy.HttpProxyBasePlugin):  # type: ignore
+    def handle_client_request(self, request: proxy.http.parser.HttpParser) -> proxy.http.parser.HttpParser:
+        # If anything fails in here, it just fails to respond
+        try:
+            with Path(self.flags.request_logfile).open("a") as f:
+                record = {
+                    "method": request.method.decode(),
+                    "url": str(request._url),
+                }
+
+                if request.headers:
+                    record["headers"] = {k2.decode().lower(): v.decode() for _, (k2, v) in request.headers.items()}
+                print("Request:")
+                print(record)
+                print()
+                f.write(f"{json.dumps(record)}\n")
+        except Exception as e:
+            # Making sure there is some visible output
+            print(repr(e))
+            traceback.print_exception(e)
+            raise e
+        return request
